# [GAPRINDASHVILI] Use AREL for custom_attributes virtual_attributes in MiqReport ONLY