@NickLaMuro 2018-06-25
Don't Merge!

Looks like I have one thing to fix causing the specs to fail.  Will get to it shortly.