@simaishi 2018-06-26
Fine backport details:

```
$ git log -1
commit cb5a8cb8cf3a9365c4b2098192a3e86a0f2d2a72
Author: Satoe Imaishi <simaishi@redhat.com>
Date:   Tue Jun 26 08:43:07 2018 -0400

    Merge pull request #17629 from NickLaMuro/custom_attributes_into_extra_cols_for_miq_reports_gapindashvili
    
    [GAPRINDASHVILI] Use AREL for custom_attributes virtual_attributes in MiqReport ONLY
    (cherry picked from commit d7d0f96a5e4cf27f093116e8f6982de220691483)
    
    Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1594028
```